[Keyless Common] Split encoding.rs into two separate files.
