+++ keyless-common/src/input_processing/mod.rs
@@ -4,4 +4,5 @@ pub mod bits;
 pub mod circuit_config;
 pub mod circuit_input_signals;
 pub mod encoding;
+pub mod jwt;
 pub mod sha;
